Move executive to manager.h
